Use Database::requireRecord when possible
If we expect a query to have exactly one record, then `requireRecord`
is more robust than using `nextRecord` in a conditional. Specifically,
it allows us to ensure that the variables that we define from the SQL
record will always exist. Even if `nextRecord` would always be true in
practice, a linter would not have that knowledge.

Note: A `count(*)` query always returns a record, so this is why we
can use `requireRecord` in some places.