Fix uninitialized arrays
These variables are appended to inside for-loops without first being
initialized, and then used outside the for-loops. If the for-loop is
an empty loop, then we would get an undefined variable warning.
Normally, we do not expect an empty loop, but this is not guaranteed,
and a good linter will warn us about this possibility.

Initializing the arrays prior to the for-loop makes appending safer
(e.g. what if the variable name was inadvertently used earlier?) and
ensures that the usage outside the for-loop is valid even in the case
of an empty loop.

Fixes phpstan warnings like:

> Variable $X might not be defined.