Replace if-statements with the PHP8 match function
When we are checking the value of some variable against an exhaustive
list of all its possible values (and if we want to raise an internal
error if we get an unexpected value), then the PHP8 `match` function
is a more robust solution than if-elseif-else statements.

The `match` construction makes it clear that the return variable will
always be set if execution continues, and thus there is no uncertainty
about whether or not the variable will exist.

It also has an advertised feature where we can match on `true`, which
allows us to branch on non-identity expressions. We use this in a few
places where some of the branches are used for a range of values.