Don't use if-else with error forwarding
If we are forwarding to an error page as the result of an if-statement,
then we don't need an else-branch, because anything we put in that
branch would work in the exact same way if put after the conditional.

When a variable was defined in such an else-branch, it would not be
clear to a linter that it would always exist (since there's not yet a
way to hint that a function such as `create_error` does not return).
Therefore, moving it outside of the conditional makes it unambiguous
without changing the behavior.