Remove some unnecessary object updates
In most cases, explicitly calling update methods within processing
pages is unnecessary, because we will always have a sector lock and
so the updates will be called automatically as part of `do_voodoo`.

This also fixes some PHPStan warnings since some of these update
methods only exist in the derived class, but we're using the base
class object.