admin_message_send_processing.php: fix type issue
Round to an integer when converting the input expire hours to seconds,
since `SmrPlayer::sendMessageFromAdmin` now requires an int argument.

We can also remove the check on negative expiration time, because
a) it is HTML5 validated on the input form and b) if the HTML5
validation is circumvented, it will raise an exception when attempting
to insert a negative value into an unsigned mysql column.