ChessGame: simplify function getColourForAccountID
This function no longer returns `false`, indicating that the given
ID is not a player of the game. Instead, it uses the match function
and _requires_ the given account ID is for one of the players.

A new method `isPlayer` now exists to fulfill the old role, which
allows us to query if the given account ID is for one of the players.