Add unit test for Template class
In order to test a private method of the Template class, we needed to
add a utility function `TestUtils::getPrivateMethod`, which uses the
PHP reflection tools to change the method visibility. Ideally, the
solution is to refactor the Template class; however, in the meantime
we still want to be able to test it thoroughly.

Also fixed a long-standing bug with the `checkDisableAJAX` method,
which was written with the expectation that it was disabling AJAX if
an input element was found that was *not* of type 'submit', 'hidden',
or 'image' (since these types don't involve any user input). Instead,
_all_ input types were disabling AJAX updates. The regex was modified
to correctly capture the original intention.