SmrPlayer: fix serialization for combat logs
We need the `npc` property included in the serialization of the
SmrPlayer class so that the "[NPC]" string can be properly suffixed
to NPC players' names in the combat log.

We initialize the property `$npc = false` so that existing combat
logs do not throw an exception. This is needed because the SmrPlayer
serializations in those old combat logs don't have the `npc` property,
and so were setting the `npc` property to null. Now that we are using
typed properties in the SmrPlayer class, we cannot set a bool to null.