AbstractSmrCombatWeapon: refactor getMaxDamage
Use the implementation of `getMaxDamage` from the SmrWeapon derived
class (i.e. dynamically calculate the max of shield/armour damage)
instead of hard-coding the max damage in the SmrMines, SmrScoutDrones,
and SmrCombatDrones classes.

This also is necessary for deserializing the SmrWeapon class in combat
logs, since that class was not initializing the `maxDamage` property.