AbstractSmrCombatWeapon: improve interface
Make the following methods abstract:

* getName
* getBaseAccuracy
* getShieldDamage
* getArmourDamage

These were previously implemented in the abstract base class; however,
they were overridden in the SmrWeapon class. Since the implementation
in the base class only made sense for the SmrMines, SmrCombatDrones,
and SmrScoutDrones classes, we move those implementations to a trait
that is used only in those classes.

This does two important things:

1. The SmrWeapon methods are no longer overrides; they implement the
   abstract methods in the base class.

2. We do not leave the `name`, `accuracy`, `shieldDamage`, and
   `armourDamage` properties uninitialized in the SmrWeapon class
   (which was causing the deserialization of the class in combat logs
   to fail due to trying to set these typed properties to null).