Define ship class names in PHP instead of database
The database should be reserved for data that changes over time,
especially because database queries are the slowest way to retrieve
data (whereas loading a PHP class is significantly faster).

It is also more robust to define the ship class constants entirely
in PHP, since we still needed to define the numerical values in PHP
even when we were using a database.

(NOTE: In PHP 8.1, this class could be converted to an enum.)