Define user rankings in PHP instead of database
The database should be reserved for data that changes over time,
especially because database queries are the slowest way to retrieve
data (whereas loading a PHP class is significantly faster).

I have tried to consolidate all the logic for user rankings in this
class; however, the *per stat* ranking details remain in the
AbstractSmrAccount class.