npc.php: fix uninitialized session var
We need to be able to get the session var before we process any pages,
but it is during page processing where we set the session var. Thus,
we need to initialize the session var to an empty Page (as a kind of
bootstrapping strategy).

This fixes the following error:

> Error Message: ErrorException: Undefined array key "" in /smr/src/lib/Smr/Session.php:259
> Stack trace:
> #0 /smr/src/lib/Smr/Session.php(259): exception_error_handler(2, 'Undefined array...', '/smr/src/lib/Sm...', 259)
> #1 /smr/src/lib/Default/AbstractSmrPlayer.class.php(2169): Smr\Session->getCurrentVar()
> #2 /smr/src/tools/npc/npc.php(140): AbstractSmrPlayer->removeUnderAttack()
> #3 /smr/src/tools/npc/npc.php(91): NPCStuff()
> #4 {main}