npc.php: fix missing var needed to UNO
The shop_hardware_processing.php script forwards the current LocationID
in the session var to the next page. This was not being set as part of
the preparation in npc.php (because the LocationID is not needed by
anything the NPC does, i.e. no subsequent display page). However, it is
no significant burden to provide this, so we pass along the sector ID
to the `doUNO` function.

Fixes the following error:

> Error Message: Exception: Could not find "LocationID" in var! in /smr/src/lib/Default/Page.class.php:166
> Stack trace:
> #0 /smr/src/engine/Default/shop_hardware_processing.php(64): Page->addVar('LocationID')
> #1 /smr/src/lib/Default/Page.class.php(231): require('/smr/src/engine...')
> #2 /smr/src/lib/Default/smr.inc.php(372): Page->process()
> #3 /smr/src/tools/npc/npc.php(352): do_voodoo()
> #4 /smr/src/tools/npc/npc.php(176): processContainer(Object(Page))
> #5 /smr/src/tools/npc/npc.php(91): NPCStuff()
> #6 {main}