npc.php: fix session init in db logging mode
When logging debug messages to the database (as we do in beta), we
were using the session prior to initializing the session var. As a
result, we were running into the same issue as in a15d403a972e. The
fix is to initialize the session var prior to any debug statements.