Use discord reply instead of sendMessage
In DiscordPHP v6, the `reply` method on a Message uses the new Discord
reply mechanism. Previously, it just @'d the user, which I didn't like,
which is why we mostly used the `sendMessage` method on the Channel.

The new reply mechanism is very nice though, so we replace all uses of
`sendMessage` with `reply`.

Two additional changes to the GameLink class:

1. Accept a Message instance in the ctor instead of its Channel and
   User components. This both simplifies the interface and allows us
   to use the `reply` method.

2. Demote `$alliance` and `$account` from class attributes to local
   variables. These are not used outside of the GameLink class, and
   even if they were we could get them from the `$player` anyway.