Fix missing check for excess when buying a new ship
We need to add a call to `checkForExcess` in `setShipTypeID`. Without
this, the excess hardware/weapons/cargo are not removed when buying a
new ship.

This was broken by c2cffc5294, where `checkForExcess` was removed from
`regenerateBaseShip`, which is called in both `setShipTypeID` and the
ctor. However, the excess check must only be done in `setShipTypeID`,
since hardware/weapons/cargo aren't even loaded in the ctor.