SmrLocation: avoid db calls in HunterWars child
To avoid (mostly) redundant database calls, we call the parent
implementation of this method first, and then remove ships from the
list that don't belong in a Hunter Wars game.

The removal of the database call if worth the trivial extra expense
of constructing a few extra `SmrShipType` objects (since we don't
incur any additional database queries).