SmrSector: fix setGalaxyID for new sectors
When property types were added to SmrSector in 456b7c6d, we broke the
`setGalaxyID` function because it was trying to access the `galaxyID`
property before it was set. We now perform an `isset` check to avoid
this issue (which is probably preferable to initializing the `galaxyID`
to some invalid value, e.g. 0).