Fix improper LEFT JOIN usage
When using "LEFT JOIN" between tables like `planet`, `port`, etc.
and `sector`, we were inadvertently returning a database record for
every sector (instead of just the ones that have a planet/port/etc.).
This was because we put the `sector` table first. Switching the order
dramatically reduces the number of returned records.

There was special logic in many of the `getGalaxy{Planet,Port,...}`
functions to deal with non-existent records, which can be removed as
a result of this fix (since only existing records will be returned).

Note: Alternatively, instead of a "LEFT JOIN", we could have used an
"INNER JOIN" (or just "JOIN"), which would furthermore omit records
for which the planet/port/etc. exist, but the associated sector does
not. However, I think that it's perhaps useful to return all of the
records so that errors are raised for invalid ones.