DatabaseRecord: add getString method
The `getString` method should be used when the returned value can
_only_ be a string. (If it can be null, use `getField` instead.)

This signals to static analyzers that the returned value will be
compatible with the string-only context in which it is subsequently
used. (Though in reality it really just shifts the point of error for
a non-string return value from the assumed string-only context in
which it's used to the `getString` function, where a non-string value
will raise a TypeError exception.)

Fixes warnings of the following type:

> It seems like $dbRecord->getField('location') can also be of type
> null; however, parameter $string of stripslashes() does only seem
> to accept string, maybe add an additional type check?

In cases where we change from `getField` to `getString`, I verified
that the database column has "NOT NULL".

In other cases, the warnings were fixed by removing the string-only
context (e.g. removing a call to `stripslashes` when not needed,
especially since our database escaping no longer adds slashes).