shared.css: fix lmloc div placement
Each Local Map ("lm") sector is 120px wide, but we need to leave 5px
on each side for the walls. Therefore, the `lmloc`-classed divs need
to be shortened in width by 10px so that, e.g. we do not overlap with
the wall add/remove link in the Uni Gen.