SmrSector: fix initialization of links
Only populate the `links` property with real connections (i.e. do not
add connections to sector 0, since that indicates no link exists in a
particular direction).

This allows us to simplify some functions (both inside and outside the
SmrSector class) that were checking if any link sector returned by
`getLinks` was 0. Now this will never be the case.

The initialization of `links` was intentionally changed in 7054e3d9 to
include all directions, but it did not intend the cascade of problems
that this would create (e.g. NPC's failing to find sector 0, writing
connections to sector 0 in the SMR Sectors file, etc.).