npc.php: never continue from action loop
Since we do all the appropriate state cleanup when processing a caught
ForwardException, make sure that we never naturally continue from the
NPC action loop. That means replacing `continue` and throwing a fatal
error if we reach the end of the loop.

As a result, we can remove the ad-hoc `$_REQUEST` cleanup at the top
of the loop.