Uni Gen: add ability to drag and drop locations
This new feature allows us to move locations and planets to different
sectors in the Uni Gen (instead of the laborious task of going into
the details page for one sector, removing the location, then going to
the details of another sector and adding that location). The Uni Gen
should be a much more pleasant experience with this change.

The implementation is achieved with jQuery UI Draggable and Droppable
widgets. We attach Draggable handlers to all `.drag_loc` CSS classes
(defined only in the Uni Gen) and Droppable handlers to `.lm_sector`,
so that any location can be dragged into any sector box.

Any change in the DOM requires us to rebuild these handlers, and so
we add the `setupDragDrop` function as a callback to the location move
event and to the sector wall toggle event.

Note that the jQuery UI handlers do not get properly rebuilt during
normal AJAX refreshes if the Uni Gen map changes remotely (i.e. if it
is updated by another admin). This could be fixed by installing the
following document listener:

  $(document).ajaxComplete(function(event, request, settings) {
      setupDragDrop();
  });

However, it is an expensive operation, which will be unneeded most of
the time (it is not clear to me if it does anything if there are no
changes in the DOM).