Update to phpDocumentor 3.1.0
This version adds support for PHP8, so we update the Docker image from
PHP 7.4.14 to 8.0.7 (to match the game image).

Grab the phpDocumentor.phar file from the GitHub releases, since that
allows us to download a specific version (whereas phpdoc.org provides
only the latest version). GitHub is also likely to be a more reliable
endpoint.