SmrLocation: fix array element ordering in cache
When we move a location using `moveSectorLocation`, it will be added
to the _end_ of the `$CACHE_SECTOR_LOCATIONS` array, even though it
is indexed by location type ID (that's just how PHP associative arrays
work).

The ordering of the cache is different when populated from the
`getSectorLocations` method, which results in the Uni Gen locations
potentially changing their order on the first refresh after a loc has
been moved. Since this is a DOM update caused by an AJAX refresh, it
was breaking the drag/drop handlers attached to those elements.

To fix this, we simply sort the cache array by key after calling
`moveSectorLocation` to get the expected type ID based ordering.