Replace race table with Race class
Since the race properties are constants, it is more performant to keep
that information in a PHP class instead of the database.

The race functions in the Globals class have been moved to the new Race
(and RaceDetails) class. Some of the API changes include:

* Globals::getRaces -> Race::getAllNames
* array_keys(Globals::getRaces) -> Race::getAllIDs
* Globals::getRaceName -> Race::getName
* Race::getPlayableIDs (new method that excludes RACE_NEUTRAL)

All racial descriptions have been updated. The Ik'Thorne and Creonti
images have been swapped to fit their descriptions better.