Session: remove updateVar
With the change in `$var` from an array to a `Page` (i.e. a wrapper
around `ArrayObject`), modifying `$var` directly persists in the
Session. Since it is more straightforward to modify `$var` directly,
we can remove the `SmrSession::updateVar` method.

Note that while removing `updateVar`, some instances where we didn't
need to be updating the `$var` were modified accordingly.