SmrAlliance: remove unnecessary ctor logic
We do not need to explicitly set the `kills` and `deaths` properties
to 0 under any circumstances due to both the default value for these
columns in the `alliance` table (i.e. 0) and due to the behavior of
`DatabaseRecord::getInt`, which would convert any empty-like value
(if there could be one) to 0.