Refactor rankings
The SQL queries to compute rankings were complicated for two reasons:

1. We included the logic for selecting subsets of the rankings into
   the query itself.

2. We had a separate query to get the rank associated with the current
   player, which was non-trivial to keep in sync.

We can resolve both of these issues by simply calculating all of the
ranks up front, and then using PHP manipulations to extract subsets
and to identify the current player's rank.

One might think this would be a performance degradation, because we
are not limiting the query (it has a lot more returned rows in most
cases). However, we were already doing the same amount of work in the
subquery of the query to get the current player's rank. Therefore, it
is actually a performance improvement since we are performing fewer
queries overall.

Additional changes:

* We can now get the total number of ranks by looking at the size of
  the returned result, i.e. `count($rankedStats)`, instead of using a
  separate query to get the total number of players, alliances, etc.

* We add the function `Rankings::raceStats` to avoid duplicating the
  complicated SQL query used in each of the race ranking pages.

* The sector ranking page has been rewritten to follow the pattern of
  the other ranking pages.