Fix reloading session var from database after lock
After we acquire the sector lock in `do_voodoo`, we reload the session
var from the database (even though we try to prevent spam clicking, it
can happen I suppose?).

In 3c0847f7b2fe, the session logic was changed so that the current var
was updated in-place (i.e. `getCurrentVar` would not need to be called
again to apply these changes to existing `$var` instances), and so the
`getCurrentVar` call was removed from the session var reloading logic
in `do_voodoo`.

However, this change was reverted in 5c7f65b41, but `getCurrentVar`
was not restored. As a result, we could get into an inconsistent state
in the case where the session var in the database changed between when
we first initialized the session and when we acquired the sector lock.

The fix is simply to re-add the call to `getCurrentVar` after the var
is reloaded from the database.