Fix bug sharing maps when all sectors are explored
This was a result of the `$unvisitedSectors` array being empty, which
caused the SQL query to use a `AND sector_id NOT IN ()` clause. This
clause is now omitted if the array is empty.

I suspect that this worked in the past due to the escaping mechanism
in `escapeArray`, likely resulting in `AND sector_id NOT IN ('')`
(see the `$autoQuotes` option, which was removed in 3b996d91fb5c4).