Fix error displaying scout messages
In the `displayGrouped` and `displayMessage` functions, we now return
the constructed message (an array) instead of passing in the message
box by reference and appending to it inside the functions.

This fixes an issue with scout messages where the "GroupedMessages"
element did not yet exist, but we were passing it by reference to the
`displayMessage` function as if it were an array. This caused the
following error:

> TypeError: displayMessage(): Argument #1 ($messageBox) must be of type array, null given