SmrPort: remove null return of getGoodTransaction
The SmrPort interface is more robust without a possible null return
value for `getGoodTransaction`. An exception is now thrown if called
with a goodID that the port does not buy or sell.

If you want to determine *if* a port buys or sells a good, use the
function `SmrPort::hasGood` instead (which has an optional transaction
type argument to validate against).

This fixes warnings related to using the return value of the function
`getGoodTransaction` without checking if it is null, e.g.:

> It seems like $transaction can also be of type null; however,
> parameter $transactionType of AbstractSmrPort::getIdealPrice() does
> only seem to accept string, maybe add an additional type check?