Database: do not store DatabaseProperties
We want to store the `DatabaseProperties` in as few places as possible
to avoid leaking sensitive credentials (i.e. the database password,
which was sometimes being exposed in bug reports). Since the `Database`
class really only needs the database name, we can just pass that into
the constructor instead.

Define a DI constant `DatabaseName`, which extracts the database name
from the `DatabaseProperties` in the DI container. Then we wire this
DI constant into the ctor of the `Database` class.

Note that `Database::getDbBytes` will now get the size of the current
database instead of always using the live database. This change is not
required, but it seems like a sensible thing to do.