Prevent looting a port after it resets
If you load the loot screen before a port resets and then loot it
after the reset, you can take the credits from the reset port, which
is basically a "free" PR (and I think it would leave the port empty
of credits for the next PR). It's not clear if you needed to spam the
loot button to exploit this.

Now you will get an error message if you try to loot the port after
it has reset.