Refactor PHP-DI definitions compilation
When initializing the DI container, pass a bool indicating whether or
not compilation should be enabled. This makes it much easier to test
the container, since it no longer depends on some global state of the
environment (and thus we don't have to worry about preparing and then
cleaning the environment in the tests).

* Fix definition in docker-compose.yml, which was passing in the value
  `"\"true\""` (i.e. the literal string `"true"`) instead of `"true"`
  (i.e. the literal string `true`).

* Make sure compilation is disabled in the PHPUnit infrastructure in
  all places _except_ for where we test that compilation can be
  properly enabled.

* We use `getenv(...)` instead of `$_ENV[...]` because `$_ENV` is only
  populated if the php.ini `variables_order` setting includes "E".
  It turns out that, at least in the development image, this is set to
  "GPCS", so `$_ENV` was _not_ being populated, and we therefore were
  not respecting the `DISABLE_PHPDI_COMPILATION` env var!