Highlight last chess move server-side (#1147)
We had a complicated scheme for highlighting the last move of a chess
game, where the JavaScript code would be transmitted to the browser
via ajax, and then a special "EVAL" hook would run that code.

A simpler scheme would be to put the highlighting code in the JS file,
and then only transmit the "lastMove" data via ajax.

However, the simplest scheme seems to be to just add the `lastMove`
HTML class attribute when processing the page server-side. Otherwise,
the browser needs to run the highlighting code on every ajax update,
even though most of the time the page won't be changing. This way we
don't even need to transmit the "lastMove" data to the browser at all.

Additional details:

* To pass the `lastMove` HTML class attribute via ajax, we had to move
  it to a new child `<div>` of the `<td>` element (which is the ajax
  element). This required adjusting the CSS a bit, as we now style the
  div instead of the td.

* We no longer put a `&nbsp;` in empty squares, as it does not seem to
  be necessary.