Remove resetCache methods (#1155)
These methods had two major flaws:

1. They do not actually update existing instances of classes. So any
   place where an instance was stored outside the cache could be out
   of sync with the version in the cache.

2. They do more work than simply using `clearCache` since it forces
   all existing instances to be reloaded immediately, rather than
   letting them be reloaded on demand.

The only place any of these were used was in vote_link_callback.php,
and it was being used incorrectly there. Specifically, `$player` was a
reference to the _original_ instance, prior to getting a sector lock,
and we were not updating it after calling `resetCache`. Therefore, we
were potentially overwriting data if the database was updated between
the initial `getPlayer` call and acquiring the lock. (Note that this
was not a _practical_ issue, due to requiring precise race conditions,
but a problematic design nonetheless.)