Remove delayed planet defense updates
Since the planet is always the defender in battle (i.e. fires second),
this feature was intended to allow planets to fire back on attackers
with their full strength (specifically, with the CDs they had prior to
taking damage).

This was only ever enabled between Sept-Nov 2009. It was disabled in
a1c3c81e91, referring to some "reduced damage fix" that rendered it
obsolete. It's not clear if this was ever enabled on the live server,
though.

Delayed updates only served to complicate the SmrPlanet interface and
to make the API inconsistent from other combat entities.