Handle UserException at the handler level
We were manually catching `UserException` every time that it was
thrown and immediately calling `create_error`. Since the only time we
actually want to call `create_error` is when user-facing code is being
executed, we can instead do this in the `handleException` function.

This creates a more natural flow for exception throwing/handling, and
opens up the possibility of more globally replacing `create_error`
calls (which are quite complex) with throwing `UserException`. Unit
testing would especially benefit from this.