Do not call create_error directly in libraries
Instead, throw a `UserError` exception. This is possible due to the
change to `handleException` that calls `create_error` when handling a
`UserError` exception.

The reason we want to avoid calling `create_error` directly in library
functions (and, perhaps eventually, all code) is because we can easily
unit test that a function raises an exception, but it is hard to test
that it calls `create_error` (it probably would need to be mocked,
since it is a complicated function with page forwarding logic).