Convert UserError to regular Exception
These specific errors indicate a code logic bug, not user error, so
they should raise a regular `Exception` (which will be logged) rather
than a `UserError` (which would _not_ be logged).

Generally speaking, if we're asking the user to report the error to an
admin, it should not be a `UserError`.