Simplify Blackjack\Hand interface
The `Hand` class was difficult to test because the only way to add
a `Card` was by drawing a card from a `Deck`, but the `Dec::drawCard`
method involves randomness, so you couldn't just construct an
arbitrary hand of cards.

We replace `Hand::drawCard` with `Hand::addCard`, which takes a `Card`
instead of a `Deck`.