npc.php: refactor how we switch between NPCs
Throw a `FinalActionException` instead of calling `changeNPCLogin`
directly when we are done with the current NPC. Then the decision
about whether to change NPCs or exit is handled in the catch-clause
for this exception.

The check for whether any NPCs exist at all is now in `changeNPCLogin`
just before we populate the list of available NPCs. This is a more
natural place for this check.

This is a first step in reorganizing the NPC code, which would benefit
greatly from a more object-oriented structure.