npc.php: simplify ship upgrading
Similar to the UNO/weapons change, ship upgrading will no longer use
the same machinery as human players. Instead, if an NPC has the cash
to upgrade, it will do so prior to starting.

For the future, this change will facilitate NPCs changing ships w/out
needing to have access to the shop that sells it (or meeting any other
restrictions that may complicate this procedure).