npc.php: expand ship upgrade paths
Each race now has 5 ships that they can upgrade through, concluding
with the top racial. This will achieve several objectives:

1. Create a more interesting landscape of ships.
2. Provide a challenge if an NPC is given the opportunity to fully
   upgrade (players wanted greater risk in NPC hunting).
3. Slow down NPC credit/exp growth once it passes the racial trader
   (to prevent any one NPC kill from being _too_ lucrative).

It is unnatural to have ships trading in fighting ships, and the end
goal may be better achieved by something like NPC sentries. But for
now, this may be the most efficient solution.