Fix pma service for beta.smrealms.de
The `pma` service on einstein was supposed to be exposed on the
`beta.smrealms.de/pma` route, but was not due to having lower priority
than the `smr-dev` route when the `SMR_RULE` environment variable is
set. We can reuse `SMR_RULE` for the `pma` service to increase its
priority and to ensure that it can't override `www.smrealms.de/pma`.

In the local development environment case, where we are not setting
`SMR_RULE`, we default to `PathPrefix(/)`, which is the closest we
can get to an "empty" rule in Traefik.