Dockerfile: update dependencies
* Xdebug 3.1.1 -> 3.1.2 (https://xdebug.org/updates#x_3_1_2)
* Composer 2.1.14 -> 2.2.1

Composer 2.2.0 introduces the `COMPOSER_NO_DEV` environment variable,
which is equivalent to using `--no-dev` if set to 1. We can use this
to simplify the way that we call Composer. We change our env variable
`PHP_DEBUG` to `NO_DEV` to avoid needing to invert the value. Default
is still "production".