Move NHA create to game enable
Create the NHA when the game is enabled instead of when the game is
created. This will prevent unnecessary database spam for games that
will not be enabled (e.g. maps in beta/local dev).

Also simplify the NHA alliance database insertion by setting more of
the alliance properties with setter methods.