Create NHA by name instead of by ID
If we reserve the NHA name ("Newbie Help Alliance") instead of its ID,
then we can use SmrAlliance::createAlliance to create the NHA, just as
we would for any other alliance. An exception is thrown if a player
tries to create an alliance using this name.

This means that the NHA alliance ID will be arbitrary (though likely
to have `alliance_id=1` since it is now created when the game is first
enabled). As a result, we have to change all the special cases for the
NHA to use its name instead of its ID. We add the SmrAlliance::isNHA
method to help isolate this logic.

Additional change: The 2nd arg of SmrAlliance::setAllianceDescription
is now optional, and if unused the description is not logged in the
admin box. This makes it possible to use this method when creating the
NHA, since no players exist at the time of its creation.