Refactor accountID into VoteSite class
An account ID is used in almost every method of this class, so let's
just make it a class property. This significantly simplifies the class
interface.

Also don't bother caching the VoteSite instances themselves, since
they are trivially inexpensive to construct (and are only constructed
a couple times in any given page at most).