Add indentation for refactoring
All non-class files in `lib/` are being moved into classes. To do this
requires adding a layer of indentation for the class scope.

We add the indentation separately from the functional changes to avoid
breaking git rename detection due to the combination of a file name
change and massive indentation changes.