Load smr.inc.php during bootstrapping
The `smr.inc.php` file has common functions that are used all over the
place, and will be less straightforward to convert to a single class.
As a result, it is easiest to load it once (with an explicit path) in
bootstrap.php.

In the future, this file should be broken up into several classes so
that we can load each component as needed (via the class autoloader).