Move admin to subdirectory of engine
When specifying paths to admin engine files, we must now use the
`admin/` prefix. This is more prescriptive, helps prevent collisions,
and makes the file searching functions both simpler and faster
(since all engine pages are now in the `engine/` directory).

Renamed the `admin/1.6` directory to `admin/unigen` for clarity.