bootstrap.php: fix exception_error_handler
This function is passed to `set_error_handler`, which means that it is
called for all diagnostics, regardless of the error level that is set.
Therefore, we need to account for the error level to make sure that we
properly suppress any diagnostics that we have configured to be
suppressed.

In this case, `exception_error_handler` was throwing for warnings that
we had explicitly suppressed with the "@" operator (due to a lack of
support for PHP 8.1 in a few 3rd party libraries). On the beta server,
we don't want this to raise an exception until we remove the "@".

See https://www.php.net/manual/en/language.operators.errorcontrol.php
for a reference to this design.